Extract cooldown/floor logic into timerCommandCooldowns.ts
Splitting pickRowsToFire into two functions resolved qlty's
function-complexity flag but not the file-level total, since it was
still all in one file. Moves the per-command Shared Chat cooldown and
cross-command channel floor (state, pruning, pickRowsToFire,
releaseReservations) into their own module, mirroring the earlier
timerCommandFireGate.ts extraction. The longest-waiting tie-break now
takes a lastFiredAtOf lookup instead of closing over the scheduler's
timerState map directly, so the module has no dependency on the
scheduler's internal state shape.

No behavior change — existing tests in timerCommandScheduler.test.ts
already exercise this logic thoroughly through the public
runTimerCommandTick API.
