+++ src/twitch/timers/timerCommandScheduler.ts
@@ -7,6 +7,7 @@ import {
   evaluateFireBlock, logBlockReasonChange, forgetBlockReason, pruneBlockReasonLog, clearBlockReasonLog,
   type TimerRuntimeState,
 } from './timerCommandFireGate';
+import { pickRowsToFire, releaseReservations, pruneStaleCooldowns, clearCooldowns } from './timerCommandCooldowns';
 
 const log = createLogger('TimerCommandScheduler');
 
@@ -42,82 +43,6 @@ function rowKey(row: { id: number; channel: string }): string {
 }
 
 const timerState = new Map<string, TimerRuntimeState>();
-
-/** Picks whichever of `rows` has gone longest without firing (oldest `timerState.lastFiredAt`), breaking ties by keeping the first. */
-function pickLongestWaiting(rows: readonly TimerCommandForScheduler[]): TimerCommandForScheduler {
-  return rows.reduce((oldest, row) => {
-    const oldestState = timerState.get(rowKey(oldest))!;
-    const rowState = timerState.get(rowKey(row))!;
-    return rowState.lastFiredAt < oldestState.lastFiredAt ? row : oldest;
-  });
-}
-
-// ─── Per-command Shared Chat cooldown ────────────────────────────────────────
-//
-// When several of one timer's assigned streamers are merged into the same Twitch Shared Chat
-// session (multitwitch), each of that timer's assigned channels otherwise keeps firing on its own
-// independent schedule — and since Shared Chat shows every participating channel's messages in one
-// merged view, that stacks up into the same command posting several times in a few minutes. This
-// caps a given timer to firing at most once per its own `interval_seconds` into a given session —
-// as if that whole session were one channel, for this command specifically — and rotates fairly
-// among the timer's channels sharing it (see `pickRowsToFire`) so each gets a turn instead of one
-// dominating. Scoped per (timer id, session id): a *different* timer assigned to the same
-// shared-chat streamers gets its own independent cooldown, sized to its own interval, and never
-// competes with this one for a turn.
-
-interface CommandSessionCooldown {
-  firedAt: number;
-  /** The interval (ms) that was actually applied when this cooldown was reserved — needed to prune it correctly later, since different timers (and a live-edited interval) can each imply a different cooldown length. */
-  cooldownMs: number;
-}
-
-const commandSessionLastFiredAt = new Map<string, CommandSessionCooldown>();
-
-/** Composite key for {@link commandSessionLastFiredAt}: one cooldown per (timer, Shared Chat session). */
-function commandSessionKey(timerId: number, sessionId: string): string {
-  return `${timerId}::${sessionId}`;
-}
-
-/** A stale command-session cooldown entry is kept around for this many multiples of its own cooldown length before being pruned, so the map doesn't grow unbounded as Shared Chat sessions come and go over long uptimes. */
-const COMMAND_SESSION_COOLDOWN_PRUNE_FACTOR = 10;
-
-/** Drops command-session cooldown entries old enough (relative to their own recorded cooldown length) that the session or assignment behind them is almost certainly gone. */
-function pruneStaleCommandSessionCooldowns(now: number): void {
-  for (const [key, entry] of commandSessionLastFiredAt) {
-    if (now - entry.firedAt > entry.cooldownMs * COMMAND_SESSION_COOLDOWN_PRUNE_FACTOR) {
-      commandSessionLastFiredAt.delete(key);
-    }
-  }
-}
-
-// ─── Cross-command channel floor ─────────────────────────────────────────────
-//
-// Independent of Shared Chat: if a single Twitch channel has several *different* timer commands
-// assigned to it, each fires on its own schedule with nothing otherwise stopping two of them from
-// landing seconds apart. This enforces a minimum gap between any two *different* commands posting
-// into the same channel, and — like the Shared Chat cooldown above — lets whichever has waited
-// longest go first. It deliberately does not slow down a single command's own cadence: the floor
-// only applies when the channel's last post came from a *different* timer than the one about to
-// post now, so a lone timer with a short interval (as low as `chk_timer_command_interval`'s 60s
-// floor) is never throttled below its own configured interval just for being alone on its channel.
-
-const CHANNEL_MIN_SPACING_MS = 120_000;
-
-interface ChannelCooldown {
-  firedAt: number;
-  timerId: number;
-}
-
-const channelLastFiredAt = new Map<string, ChannelCooldown>();
-
-const CHANNEL_MIN_SPACING_PRUNE_MS = 10 * CHANNEL_MIN_SPACING_MS;
-
-/** Drops channel-floor entries older than {@link CHANNEL_MIN_SPACING_PRUNE_MS} so the map doesn't grow unbounded as channels stop being assigned any timer over long uptimes. */
-function pruneStaleChannelCooldowns(now: number): void {
-  for (const [channel, entry] of channelLastFiredAt) {
-    if (now - entry.firedAt > CHANNEL_MIN_SPACING_PRUNE_MS) channelLastFiredAt.delete(channel);
-  }
-}
 
 let tickTimer: ReturnType<typeof setInterval> | null = null;
 let tickRunning = false;
@@ -161,135 +86,12 @@ async function resolveSessionIdsByChannel(
   return sessionIdByChannel;
 }
 
-/** A row picked to fire this tick, paired with the reservations {@link sendTimerRow} must release if the send doesn't succeed. */
-interface PickedRow {
-  row: TimerCommandForScheduler;
-  /** The {@link commandSessionKey} reserved for this pick, or null if its channel isn't in a Shared Chat session. */
-  sessionKey: string | null;
-}
-
-/**
- * Layer 1 of {@link pickRowsToFire}: groups `eligibleRows` by (timer id, Shared Chat session), and
- * for a group off its own cooldown (sized to that timer's `interval_seconds`), picks the row that's
- * gone longest without firing and reserves the cooldown. Rows whose channel isn't in a session pass
- * straight through untouched. See the module doc above `commandSessionLastFiredAt`.
- * @param eligibleRows - Rows that already passed their own per-timer {@link evaluateFireBlock} check.
- * @param sessionIdByChannel - Each row's channel's resolved Shared Chat session id (or null).
- * @param now - Current time in epoch ms, shared across the whole tick.
- * @returns One entry per row that cleared this layer, paired with the command-session cooldown key
- *   reserved for it (or null if its channel isn't in a session).
- */
-function applyCommandSessionCooldown(
-  eligibleRows: readonly TimerCommandForScheduler[],
-  sessionIdByChannel: ReadonlyMap<string, string | null>,
-  now: number,
-): PickedRow[] {
-  const survivors: PickedRow[] = [];
-  const bySessionAndTimer = new Map<string, TimerCommandForScheduler[]>();
-
-  for (const row of eligibleRows) {
-    const sessionId = sessionIdByChannel.get(row.channel) ?? null;
-    if (!sessionId) {
-      survivors.push({ row, sessionKey: null });
-      continue;
-    }
-    const key = commandSessionKey(row.id, sessionId);
-    const group = bySessionAndTimer.get(key);
-    if (group) group.push(row); else bySessionAndTimer.set(key, [row]);
-  }
-
-  for (const [key, rows] of bySessionAndTimer) {
-    const cooldownMs = rows[0].interval_seconds * 1000;
-    const entry = commandSessionLastFiredAt.get(key);
-    if (entry && now - entry.firedAt < cooldownMs) continue;
-
-    const picked = pickLongestWaiting(rows);
-    commandSessionLastFiredAt.set(key, { firedAt: now, cooldownMs }); // provisional — released by sendTimerRow on failure
-    survivors.push({ row: picked, sessionKey: key });
-  }
-
-  return survivors;
-}
-
-/**
- * Layer 2 of {@link pickRowsToFire}: groups whatever survived layer 1 by channel, and for a channel
- * whose last post came from a *different* timer than the longest-waiting candidate here, within
- * {@link CHANNEL_MIN_SPACING_MS}, defers it to a later tick — applied regardless of Shared Chat
- * status. See the module doc above `channelLastFiredAt`.
- * @param layer1Survivors - Rows (with their layer 1 session-key reservation, if any) that cleared {@link applyCommandSessionCooldown}.
- * @param now - Current time in epoch ms, shared across the whole tick.
- * @returns The final set of rows to actually send this tick.
- */
-function applyChannelFloor(layer1Survivors: readonly PickedRow[], now: number): PickedRow[] {
-  const toFire: PickedRow[] = [];
-  const byChannel = new Map<string, PickedRow[]>();
-  for (const picked of layer1Survivors) {
-    const group = byChannel.get(picked.row.channel);
-    if (group) group.push(picked); else byChannel.set(picked.row.channel, [picked]);
-  }
-
-  for (const [channel, picks] of byChannel) {
-    const candidate = pickLongestWaiting(picks.map((p) => p.row));
-    const candidatePick = picks.find((p) => p.row === candidate)!;
-
-    const lastPoster = channelLastFiredAt.get(channel);
-    if (lastPoster && lastPoster.timerId !== candidate.id && now - lastPoster.firedAt < CHANNEL_MIN_SPACING_MS) continue;
-
-    channelLastFiredAt.set(channel, { firedAt: now, timerId: candidate.id }); // provisional — released by sendTimerRow on failure
-    toFire.push(candidatePick);
-  }
-
-  return toFire;
-}
-
-/**
- * From the rows that already passed their own {@link evaluateFireBlock} check, decides which ones
- * actually get to fire this tick: {@link applyCommandSessionCooldown} (per-command Shared Chat
- * cooldown), then {@link applyChannelFloor} (cross-command channel floor) on whatever survives.
- * Both layers' reservations are only provisional: {@link sendTimerRow} releases them if the send
- * doesn't actually succeed, so a failed send can't silence a session/channel for the full cooldown
- * window. Rows not picked at either layer are left untouched so they're reconsidered on a later tick.
- * @param eligibleRows - Rows that already passed their own per-timer {@link evaluateFireBlock} check.
- * @param sessionIdByChannel - Each row's channel's resolved Shared Chat session id (or null).
- * @param now - Current time in epoch ms, shared across the whole tick.
- */
-function pickRowsToFire(
-  eligibleRows: readonly TimerCommandForScheduler[],
-  sessionIdByChannel: ReadonlyMap<string, string | null>,
-  now: number,
-): PickedRow[] {
-  const layer1Survivors = applyCommandSessionCooldown(eligibleRows, sessionIdByChannel, now);
-  return applyChannelFloor(layer1Survivors, now);
-}
-
-/**
- * Releases the provisional cooldown reservations {@link pickRowsToFire} made for a row, if the send
- * that would have justified them didn't actually happen (no runtime, or `runtime.send` threw).
- * Guarded by matching the reservation's own identity so a release can't clobber a newer reservation
- * made by a different row on a later tick.
- * @param sessionKey - The command-session cooldown key reserved for this row, or null.
- * @param channel - The row's Twitch channel — the channel-floor reservation to potentially release.
- * @param timerId - The row's timer id — must match the channel-floor reservation's own recorded id.
- * @param now - The epoch-ms timestamp the reservations were made under, shared across the whole tick.
- * @returns Nothing — mutates the module-level cooldown maps in place.
- */
-function releaseReservations(sessionKey: string | null, channel: string, timerId: number, now: number): void {
-  if (sessionKey) {
-    const entry = commandSessionLastFiredAt.get(sessionKey);
-    if (entry && entry.firedAt === now) commandSessionLastFiredAt.delete(sessionKey);
-  }
-  const channelEntry = channelLastFiredAt.get(channel);
-  if (channelEntry && channelEntry.firedAt === now && channelEntry.timerId === timerId) {
-    channelLastFiredAt.delete(channel);
-  }
-}
-
 /**
  * Posts one selected timer row to its channel and records the fire in the in-memory state.
  * No-ops if no runtime is registered. Never throws — a send failure is logged and swallowed so it
  * can't block other rows in the same tick. If the send doesn't succeed (no runtime, or
- * `runtime.send` throws), releases the cooldown reservation(s) {@link pickRowsToFire} made for this
- * row — see {@link releaseReservations}.
+ * `runtime.send` throws), releases the cooldown reservation(s) `pickRowsToFire` made for this row
+ * — see {@link releaseReservations}.
  * @param row - The timer's config, joined with its Twitch channel.
  * @param now - Current time in epoch ms, shared across the whole tick.
  * @param sessionKey - The command-session cooldown key this row was provisionally reserved for, or null.
@@ -323,9 +125,9 @@ async function sendTimerRow(row: TimerCommandForScheduler, now: number, sessionK
  * instead of looking identical to a healthy row still waiting out its interval), resolves Shared
  * Chat sessions for the ones that are ready, picks which of those actually get to fire this tick
  * (applying the per-command Shared Chat cooldown and the cross-command channel floor — see
- * {@link pickRowsToFire}), and sends the picked rows concurrently via `Promise.allSettled` so one
- * channel's send failure can't block another's. No-ops (re-uses the in-flight promise) if a tick
- * is already running.
+ * `pickRowsToFire` in `timerCommandCooldowns.ts`), and sends the picked rows concurrently via
+ * `Promise.allSettled` so one channel's send failure can't block another's. No-ops (re-uses the
+ * in-flight promise) if a tick is already running.
  */
 export async function runTimerCommandTick(): Promise<void> {
   if (tickRunning) return currentTickPromise;
@@ -336,8 +138,7 @@ export async function runTimerCommandTick(): Promise<void> {
       pruneStaleTimerState(new Set(rows.map(rowKey)));
 
       const now = Date.now();
-      pruneStaleCommandSessionCooldowns(now);
-      pruneStaleChannelCooldowns(now);
+      pruneStaleCooldowns(now);
 
       const eligibleRows: TimerCommandForScheduler[] = [];
       for (const row of rows) {
@@ -354,7 +155,8 @@ export async function runTimerCommandTick(): Promise<void> {
 
       const loginUserIds = timerCommandsRuntime.get()?.getLoginUserIds() ?? new Map<string, string>();
       const sessionIdByChannel = await resolveSessionIdsByChannel(eligibleRows.map((row) => row.channel), loginUserIds);
-      const toFire = pickRowsToFire(eligibleRows, sessionIdByChannel, now);
+      const lastFiredAtOf = (row: TimerCommandForScheduler): number => timerState.get(rowKey(row))!.lastFiredAt;
+      const toFire = pickRowsToFire(eligibleRows, sessionIdByChannel, now, lastFiredAtOf);
 
       await Promise.allSettled(toFire.map(({ row, sessionKey }) => sendTimerRow(row, now, sessionKey)));
     } catch (err) {
@@ -383,7 +185,6 @@ export async function stopTimerCommandScheduler(): Promise<void> {
   if (tickTimer) { clearInterval(tickTimer); tickTimer = null; }
   await currentTickPromise;
   timerState.clear();
-  commandSessionLastFiredAt.clear();
-  channelLastFiredAt.clear();
+  clearCooldowns();
   clearBlockReasonLog();
 }
